static plugin loader working
still needs to work inside extension
